removed listId from database and all pages. made primary keys for readingList table userId and bookId. #24

+++ server/index.js
@@ -105,7 +105,7 @@ app.get('/api/bookShelf/:list/:userId', (req, res, next) => {
 });
 
 app.post('/api/bookShelf/', (req, res, next) => {
-  const { title, author, bookId, coverUrl, rating, isRead, userId, listId } = req.body;
+  const { title, author, bookId, coverUrl, rating, isRead, userId } = req.body;
   if (!title || !author || !bookId) {
     throw new ClientError(401, 'invalid post');
   }
@@ -117,12 +117,12 @@ app.post('/api/bookShelf/', (req, res, next) => {
   returning *
   `;
   const readingListSql = `
-  insert into "readingList" ("title", "bookId", "rating", "isRead", "userId", "listId")
-  values ($1, $2, $3, $4, $5, $6)
+  insert into "readingList" ("title", "bookId", "rating", "isRead", "userId")
+  values ($1, $2, $3, $4, $5)
   returning *
   `;
   const bookParams = [title, author, bookId, coverUrl];
-  const listParams = [title, bookId, rating, isRead, userId, listId];
+  const listParams = [title, bookId, rating, isRead, userId];
   db.query(bookSql, bookParams)
     .then(result => {
       return db.query(readingListSql, listParams)
@@ -145,15 +145,15 @@ app.post('/api/bookShelf/', (req, res, next) => {
 
 app.patch('/api/bookShelf/:bookId', (req, res, next) => {
   const bookId = req.params.bookId;
-  const { rating, listId } = req.body;
+  const { rating, userId } = req.body;
   const sql = `
   update "readingList"
     set "rating" = $1
     where "bookId" = $2
-    and "listId" = $3
+    and "userId" = $3
     returning *
   `;
-  const params = [rating, bookId, listId];
+  const params = [rating, bookId, userId];
   db.query(sql, params)
     .then(result => {
       const [entry] = result.rows;
@@ -162,16 +162,16 @@ app.patch('/api/bookShelf/:bookId', (req, res, next) => {
     .catch(err => next(err));
 });
 
-app.delete('/api/bookShelf/:bookId/:listId', (req, res, next) => {
+app.delete('/api/bookShelf/:bookId/:userId', (req, res, next) => {
   const bookId = req.params.bookId;
-  const listId = req.params.listId;
+  const userId = req.params.userId;
   const sql = `
   delete from "readingList"
     where "bookId" = $1
-    and "listId" = $2
+    and "userId" = $2
     returning *
   `;
-  const values = [bookId, listId];
+  const values = [bookId, userId];
   db.query(sql, values)
     .then(result => {
       const book = result.rows[0];
